#199 レスポンスのカード情報部分を共通化。
