+++ app/Http/Resources/ArrivalGroupingLogResource.php
@@ -1,0 +1,31 @@
+<?php
+namespace App\Http\Resources;
+
+use App\Libs\CarbonFormatUtil;
+use Illuminate\Http\Request;
+use Illuminate\Http\Resources\Json\JsonResource;
+use App\Services\Constant\StockpileHeader as Header;
+use App\Services\Constant\CardConstant as Con;
+use App\Libs\CardInfoJsonUtil;
+
+class ArrivalGroupingLogResource extends JsonResource {
+
+    public function toArray(Request $request)
+    {
+        $array =  [
+            Con::ID => $this->id,
+            Con::ATTR => $this->attr,
+            Header::NAME => $this->cardname,
+            Header::LANG => $this->lang,
+            Header::ARRIVAL_DATE => CarbonFormatUtil::toDateString($this->arrival_date),
+            'sum_cost' => $this->sum_cost,
+            'item_count' => $this->item_count,
+            Header::VENDOR => [
+                Header::VENDOR_TYPE_ID => $this->vendor_type_id,
+                Header::NAME => $this->vendor
+            ],
+        ];
+        $array = CardInfoJsonUtil::setFoilInfo($array, $this->is_foil, $this->foiltag);
+        return $array;
+    }
+}
