+++ tests/Unit/DB/Arrival/ArrivalLogSearchTest.php
@@ -8,6 +8,7 @@ use Illuminate\Foundation\Testing\WithFaker;
 use Tests\TestCase;
 use App\Services\Constant\StockpileHeader as Header;
 use Tests\Trait\GetApiAssertions;
+use Tests\Util\TestDateUtil;
 
 use function PHPUnit\Framework\assertTrue;
 
@@ -46,7 +47,7 @@ class ArrivalLogSearchTest extends TestCase
     public function okProvider() {
         return [
             '入荷先カテゴリが買取以外' => 
-                [[Header::ARRIVAL_DATE => '2025/03/17', Header::VENDOR_TYPE_ID => 1]],
+                [[Header::ARRIVAL_DATE => TestDateUtil::formatToday(), Header::VENDOR_TYPE_ID => 1]],
         ];
     }
     /**
